+++ scripts/libuuid.sh
@@ -1,0 +1,11 @@
+#!/bin/bash
+# SPDX-License-Identifier: (LGPL-2.1 OR BSD-2-Clause)
+# Copyright (C) 2026 Rong Tao. All rights reserved.
+
+gen_uuid() {
+	if [[ -e /proc/sys/kernel/random/uuid ]]; then
+		cat /proc/sys/kernel/random/uuid
+	else
+		uuid
+	fi
+}