+++ CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to
 #### Breaking Changes
 - bpf: test gcc toolchain;
 #### Added
-- scripts: add libnet.sh;
+- scripts: add libnet.sh, libuuid.sh;
 - glibc: test bcmp() and bzero(), memccpy();
 - version: config.json: add trace-cmd;
 - bpf: add verifier subdir;